ingestion/pyproject.toml
[tool.poetry]
name = "ingestion"
version = "0.1.0"
description = "hands-on big data project: ingestion layer"
authors = ["Anzhe Meng <anzhemeng@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
